# public/openapi.yaml
openapi: 3.0.3
info:
  title: DVD Rental API
  description: DVD Rental dataset API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /auth/refresh:
    get:
      tags:
        - Authorization
      security:
        - refresh_token: [ ]
      summary: Refresh token
      description: Sets new access_token and logged_in(HTTPOnly:false). refresh_token stays the same until expires
      operationId: RefreshToken
      responses:
        '201':
          $ref: '#/components/responses/AccessToken'
        '400':
          $ref: '#/components/responses/GenericError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/GenericError'
...
components:
  securitySchemes:
    access_token:
      description: |
        You have to have access token in access token cookie to authorize request
      type: apiKey
      name: access_token
      in: cookie
    refresh_token:
      description: |
        Refresh token is used to gain new access token
      type: apiKey
      name: refresh_token
      in: cookie
    Authorization:
      type: apiKey
      name: Authorization
      in: header
      description: |
        Prefix the value with "Authorization" to indicate the custom authorization type
  responses:
...
    AccessToken:
      description: Access token
      headers:
        Set-Cookie:
          schema:
            type: string
            example: access_token httponly, logged_in boolean
      content:
        application/json:
          schema:
            type: object
            properties:
              access_token:
                type: string
                example: secureToken
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: no access token
    GenericError:
      description: String error
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: something went wrong
...